Move changii intermediates out
